exercice AssertJ and Mockito/ homework done

+++ src/main/java/com/modul_2/VatProvider.java
@@ -1,10 +1,6 @@
 package com.modul_2;
 
-import java.math.BigDecimal;
-
 public interface VatProvider {
-	
-	BigDecimal getDefaultVat();
-	
-	BigDecimal getVatForType(String type);
+	double getDefaultVat();	
+	double getVatForType(String type);
 }
